requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.10.3,<3.11.0
django-filter>=2.2.0,<2.3.0
